refactor(workhub): clarify renderer decision seams
